backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
channels==4.0.0
daphne==4.0.0
numpy==1.26.2
opencv-python==4.8.1.78
spacy==3.7.2
gTTS==2.4.0
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.2
